add debug draw function + fix type errors in create.cc

+++ src/utils/consts.hh
@@ -7,10 +7,10 @@ inline const int GAME_TARGET_WIDTH = 640;
 inline const int GAME_TARGET_HEIGHT = 480;
 
 /// @brief Tile size on screen in pixels.
-inline const int TILE_SIZE = 16;
+inline const float TILE_SIZE = 16;
 
 /// @brief Tile texture size in pixels. The original Yume Nikki draws textures at twice their size.
-inline const int TILE_TEXTURE_SIZE = TILE_SIZE / 2;
+inline const float TILE_TEXTURE_SIZE = TILE_SIZE / 2;
 
 /// @brief World chunk width in tiles.
 /// @see `World.hh`
